20250827/fin-docai-v1/requirements.txt
pdfplumber==0.11.4
pillow==10.4.0
pydantic==2.8.2
fastapi==0.112.2
uvicorn==0.30.6
python-multipart==0.0.9
numpy==1.26.4
pandas==2.2.2
streamlit==1.37.1
tabulate==0.9.0
reportlab==4.2.2
requests==2.32.3
networkx==3.3
pyvis==0.3.2
rdflib==7.0.0
streamlit-drawable-canvas==0.9.3
